[C-1] revised fix for large rewardsPerUnitTime

+++ contracts/extension/interface/IStaking1155.sol
@@ -57,10 +57,10 @@ interface IStaking1155 {
      *  @param endTimestamp       Condition end timestamp.
      */
     struct StakingCondition {
-        uint128 timeUnit;
-        uint128 rewardsPerUnitTime;
-        uint128 startTimestamp;
-        uint128 endTimestamp;
+        uint256 timeUnit;
+        uint256 rewardsPerUnitTime;
+        uint256 startTimestamp;
+        uint256 endTimestamp;
     }
 
     /**
